Make it easier to name `RetryFutureConfig` types
The purpose of `RetryFutureConfig` is to retry multiple futures in the
same way, without repeating yourself.

However making a `RetryFutureConfig` would require naming its type to
for example return it from a function. If you were using a custom "on
retry" callback that would often be impossible.

This introduces `RetryFutureConfig::boxed_on_retry` which converts the
"on retry" callback type into a `BoxOnRetryFuture<E>`. That makes the
final config type something like `RetryFutureConfig<T,
BoxOnRetryFuture<E>>` where `T` is backoff strategy.

You still have to name the type of the backoff strategy but that is only
problematic if you're using a custom strategy, which is rare. Actually
making a usable boxed `BackoffStrategy` is hard so I think we should
punt on that for now.

Example usage:

```rust
let my_retry_config: RetryFutureConfig<LinearBackoff, BoxOnRetry<io::Error>> =
    RetryFutureConfig::new(10)
        .linear_backoff(Duration::from_millis(10))
        .on_retry(|_, _, error: &io::Error| {
            // the future we return here must be 'static so we cannot capture
            // `error` as its a reference.
            let error = error.to_string();
            async move {
                println!("retrying because: {}", error);
            }
        })
        .boxed_on_retry();
```

While I was at it I made a few other minor breaking changes:

- Tokio 0.2 support has been removed. Tokio 1.x is now the only version
supported.
- `OnRetry::on_retry` no longer receives `&mut self` but instead
`&self`. That was necessary to implement `BoxOnRetry`. Users who need
mutation in their callbacks should use a mutex.
- The output of the future returned by `OnRetry::on_retry` has been
fixed to `()`. As the future is given to `tokio::spawn` I think
requiring `()` is nicer. It also simplified some trait bounds on
`BoxOnRetry`.