Make it easier reuse `RetryFutureConfig` values (#17)
The purpose of `RetryFutureConfig` is to retry multiple futures in the same way,
without repeating yourself.

However sharing a `RetryFutureConfig` would often require naming its type to for
example return it from a function. If you were using a custom "on retry"
callback that would often be impossible as the callback was required to be a
function/closure.

This changes that so the "on retry" callback can be any type that implements
`OnRetry`. So to created a `RetryFutureConfig` with a custom callback can be
done by creating a struct:

Example usage:

```rust
#[derive(Clone)]
struct MyOnRetry;

impl OnRetry<std::io::Error> for MyOnRetry {
    type Future = futures::future::BoxFuture<'static, ()>;

    fn on_retry(
        &mut self,
        attempt: u32,
        next_delay: Option<Duration>,
        previous_error: &std::io::Error,
    ) -> Self::Future {
        Box::pin(async move {
            // do something!
        })
    }
}

let config: RetryFutureConfig<LinearBackoff, MyOnRetry> = RetryFutureConfig::new(10)
    .linear_backoff(Duration::from_millis(100))
    .on_retry(MyOnRetry);
```

This can then be returned from functions or put in a `once_cell::sync::Lazy` for
sharing the config.